refactor: remove WorldProvider classes and reuse NAChunk getters for all platforms
