--- natosatlas-core/src/main/java/dev/natowb/natosatlas/core/platform/PlatformWorldProvider.java
@@ -1,14 +1,0 @@
-package dev.natowb.natosatlas.core.platform;
-
-import dev.natowb.natosatlas.core.data.*;
-
-import java.util.List;
-
-public interface PlatformWorldProvider {
-
-    NAChunk getChunk(NACoord chunkCoord);
-
-    NAChunk getChunkFromDisk(NACoord chunkCoord);
-
-    List<NARegionFile> getRegionMetadata();
-}
