+++ natosatlas-stationapi/src/main/java/dev/natowb/natosatlas/stationapi/PlatformST.java
@@ -10,7 +10,7 @@ import java.nio.file.Path;
 public class PlatformST extends Platform {
 
     public PlatformST() {
-        super(new PlatformPainterST(), new PlatformWorldProviderST(), new BlockAccessST());
+        super(new PlatformPainterST(), new BlockAccessST());
     }
