--- natosatlas-stationapi/src/main/java/dev/natowb/natosatlas/stationapi/PlatformWorldProviderST.java
@@ -1,148 +1,0 @@
-package dev.natowb.natosatlas.stationapi;
-
-import dev.natowb.natosatlas.core.NatosAtlas;
-import dev.natowb.natosatlas.core.data.*;
-import dev.natowb.natosatlas.core.platform.PlatformWorldProvider;
-import dev.natowb.natosatlas.core.utils.LogUtil;
-import dev.natowb.natosatlas.core.utils.NAPaths;
-import dev.natowb.natosatlas.core.wrapper.BlockAccess;
-import dev.natowb.natosatlas.core.wrapper.ChunkWrapper;
-import net.fabricmc.loader.api.FabricLoader;
-import net.minecraft.client.Minecraft;
-import net.minecraft.world.LightType;
-import net.minecraft.world.chunk.Chunk;
-import net.minecraft.world.chunk.storage.RegionChunkStorage;
-import net.minecraft.world.chunk.storage.RegionFile;
-
-import java.io.File;
-import java.util.*;
-import java.util.List;
-
-import static dev.natowb.natosatlas.core.utils.Constants.BLOCKS_PER_MINECRAFT_CHUNK;
-
-public class PlatformWorldProviderST implements PlatformWorldProvider {
-    Minecraft mc = (Minecraft) FabricLoader.getInstance().getGameInstance();
-
-    @Override
-    public List<NARegionFile> getRegionMetadata() {
-        List<NARegionFile> result = new ArrayList<>();
-
-        File regionDir = new File(NAPaths.getWorldSavePath().toFile(), "region");
-        File[] regionFiles = regionDir.listFiles((dir, name) -> name.endsWith(".mcr") || name.endsWith(".mca"));
-        if (regionFiles == null || regionFiles.length == 0) {
-            return result;
-        }
-
-        int index = 0;
-
-        for (File regionFile : regionFiles) {
-            index++;
-            boolean success = false;
-
-            try {
-                String name = regionFile.getName();
-                String[] parts = name.substring(2, name.length() - 4).split("\\.");
-                if (parts.length != 2) {
-                    continue;
-                }
-
-                int rx = Integer.parseInt(parts[0]);
-                int rz = Integer.parseInt(parts[1]);
-                NACoord regionCoord = new NACoord(rx, rz);
-
-                NARegionFile naRegion = new NARegionFile(regionFile, regionCoord);
-
-                RegionFile rf = new RegionFile(regionFile);
-                for (int x = 0; x < NARegionFile.CHUNKS_PER_REGION; x++) {
-                    for (int z = 0; z < NARegionFile.CHUNKS_PER_REGION; z++) {
-                        if (rf.hasChunkData(x, z)) {
-                            naRegion.chunkExists[x][z] = true;
-                        }
-                    }
-                }
-                rf.close();
-
-                result.add(naRegion);
-                success = true;
-
-            } catch (Exception ignored) {
-            }
-
-            if (success) {
-                LogUtil.info("[{}/{}] Successfully processed region file: {}", index, regionFiles.length, regionFile.getName());
-
-            } else {
-                LogUtil.info("[{}/{}] Failed to processed region file: {}", index, regionFiles.length, regionFile.getName());
-            }
-        }
-
-        return result;
-    }
-
-
-    @Override
-    public NAChunk getChunk(NACoord chunkCoord) {
-        ChunkWrapper chunk = NatosAtlas.get().getCurrentWorld().getChunk(chunkCoord);
-
-        if (chunk == null) {
-            return null;
-        }
-
-        NAChunk nac = new NAChunk();
-        for (int z = 0; z < BLOCKS_PER_MINECRAFT_CHUNK; z++) {
-            for (int x = 0; x < BLOCKS_PER_MINECRAFT_CHUNK; x++) {
-                int worldBlockX = chunkCoord.x * BLOCKS_PER_MINECRAFT_CHUNK + x;
-                int worldBlockZ = chunkCoord.z * BLOCKS_PER_MINECRAFT_CHUNK + z;
-
-                int height = chunk.getTopSolidBlockY(x, z) - 1;
-                int aboveId = chunk.getBlockId(x, height + 1, z);
-                if (BlockAccess.getInstance().isBlock(aboveId, BlockAccess.BlockIdentifier.SNOW)) {
-                    height = height + 1;
-                }
-
-                int blockId = chunk.getBlockId(x, height, z);
-                int depth = chunk.computeFluidDepth(x, height, z);
-                int blockLight = chunk.getBlockLight(x, height + 1, z);
-                int meta = chunk.getBlockMeta(x, height, z);
-                NABiome biome = NatosAtlas.get().getCurrentWorld().getBiome(NACoord.from(worldBlockX, worldBlockZ));
-
-                nac.set(x, z, height, blockId, depth, blockLight, meta, biome);
-            }
-        }
-
-        return nac;
-    }
-
-
-    @Override
-    public NAChunk getChunkFromDisk(NACoord chunkCoord) {
-        ChunkWrapper chunk = NatosAtlas.get().getCurrentWorld().getChunkFromDisk(chunkCoord);
-
-        if (chunk == null) {
-            return null;
-        }
-
-        NAChunk nac = new NAChunk();
-        for (int z = 0; z < BLOCKS_PER_MINECRAFT_CHUNK; z++) {
-            for (int x = 0; x < BLOCKS_PER_MINECRAFT_CHUNK; x++) {
-                int worldBlockX = chunkCoord.x * BLOCKS_PER_MINECRAFT_CHUNK + x;
-                int worldBlockZ = chunkCoord.z * BLOCKS_PER_MINECRAFT_CHUNK + z;
-
-                int height = chunk.getTopSolidBlockY(x, z) - 1;
-                int aboveId = chunk.getBlockId(x, height + 1, z);
-                if (BlockAccess.getInstance().isBlock(aboveId, BlockAccess.BlockIdentifier.SNOW)) {
-                    height = height + 1;
-                }
-
-                int blockId = chunk.getBlockId(x, height, z);
-                int depth = chunk.computeFluidDepth(x, height, z);
-                int blockLight = chunk.getBlockLight(x, height + 1, z);
-                int meta = chunk.getBlockMeta(x, height, z);
-                NABiome biome = NatosAtlas.get().getCurrentWorld().getBiome(NACoord.from(worldBlockX, worldBlockZ));
-
-                nac.set(x, z, height, blockId, depth, blockLight, meta, biome);
-            }
-        }
-        return nac;
-    }
-}
