+++ natosatlas-stationapi/src/main/java/dev/natowb/natosatlas/stationapi/WorldWrapperST.java
@@ -3,6 +3,8 @@ package dev.natowb.natosatlas.stationapi;
 import dev.natowb.natosatlas.core.data.NABiome;
 import dev.natowb.natosatlas.core.data.NACoord;
 import dev.natowb.natosatlas.core.data.NAEntity;
+import dev.natowb.natosatlas.core.data.NARegionFile;
+import dev.natowb.natosatlas.core.utils.LogUtil;
 import dev.natowb.natosatlas.core.utils.NAPaths;
 import dev.natowb.natosatlas.core.wrapper.ChunkWrapper;
 import dev.natowb.natosatlas.core.wrapper.WorldWrapper;
@@ -16,7 +18,9 @@ import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.storage.RegionChunkStorage;
+import net.minecraft.world.chunk.storage.RegionFile;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -162,5 +166,60 @@ public class WorldWrapperST implements WorldWrapper {
                 return ((Chunk) chunk).getLight(LightType.SKY, x, y, z);
             }
         };
+    }
+
+    @Override
+    public List<NARegionFile> getRegionFiles() {
+        List<NARegionFile> result = new ArrayList<>();
+
+        File regionDir = new File(NAPaths.getWorldSavePath().toFile(), "region");
+        File[] regionFiles = regionDir.listFiles((dir, name) -> name.endsWith(".mcr") || name.endsWith(".mca"));
+        if (regionFiles == null || regionFiles.length == 0) {
+            return result;
+        }
+
+        int index = 0;
+
+        for (File regionFile : regionFiles) {
+            index++;
+            boolean success = false;
+
+            try {
+                String name = regionFile.getName();
+                String[] parts = name.substring(2, name.length() - 4).split("\\.");
+                if (parts.length != 2) {
+                    continue;
+                }
+
+                int rx = Integer.parseInt(parts[0]);
+                int rz = Integer.parseInt(parts[1]);
+                NACoord regionCoord = new NACoord(rx, rz);
+
+                NARegionFile naRegion = new NARegionFile(regionFile, regionCoord);
+
+                RegionFile rf = new RegionFile(regionFile);
+                for (int x = 0; x < NARegionFile.CHUNKS_PER_REGION; x++) {
+                    for (int z = 0; z < NARegionFile.CHUNKS_PER_REGION; z++) {
+                        if (rf.hasChunkData(x, z)) {
+                            naRegion.chunkExists[x][z] = true;
+                        }
+                    }
+                }
+                rf.close();
+
+                result.add(naRegion);
+                success = true;
+
+            } catch (Exception ignored) {
+            }
+
+            if (success) {
+                LogUtil.info("[{}/{}] Successfully processed region file: {}", index, regionFiles.length, regionFile.getName());
+
+            } else {
+                LogUtil.info("[{}/{}] Failed to processed region file: {}", index, regionFiles.length, regionFile.getName());
+            }
+        }
+        return result;
     }
 }
